MongoDB Server established
Added a key and password to the MongoDB Cloud Server. Established a document for it to call to. 
Added a test database for it to call forth when used. 
Added code to close server after code is run (This way we don't waste potential memory, or accidentally send incorrect data to the server.